Factor out common part of wxListBoxItem and wxCheckListBoxItem
No real changes, this is just a refactoring extracting common code in
src/msw/listbox.cpp and src/msw/checklst.cpp into a new header which can
be reused from both places.

As a side effect, wxCheckListBoxItem now uses the wxListBox fix for
background colour from the parent commit, which makes it unnecessary to
set it explicitly in it any more.

Also test customizing the text colour of wxCheckListBoxItem in the
sample for completeness.
